Leason 3

+++ Matchismo/Matchismo/Matchismo/CardGameViewController.m
@@ -55,7 +55,7 @@
         [cardButton setTitle:[self titleForCard:card] forState:UIControlStateNormal];
         [cardButton setBackgroundImage:[self backgroundImageForCard:card] forState:UIControlStateNormal];
         cardButton.enabled=!card.matched;
-        self.scoreLabel.text=[NSString stringWithFormat:@"Score: %ld",self.game.score];
+        self.scoreLabel.text=[NSString stringWithFormat:@"Score: %ld",(long)self.game.score];
         
     }
 }
